+++ ui/pages/DataManagement/DataManagement.jsx
@@ -7,7 +7,9 @@ import { getUser, getElasticsearchEnabled } from 'redux/selectors'
 import { Error404, Error401 } from 'shared/components/page/Errors'
 import { SimplePageHeader } from 'shared/components/page/PageHeaderLayout'
 
+import AddIGV from './components/AddIGV'
 import ElasticsearchStatus from './components/ElasticsearchStatus'
+import LoadData from './components/LoadData'
 import RnaSeq from './components/RnaSeq'
 import SampleQc from './components/SampleQc'
 import Users from './components/Users'
@@ -16,7 +18,13 @@ import WritePedigree from './components/WritePedigree'
 
 const IFRAME_STYLE = { position: 'fixed', left: '0', top: '95px' }
 
+const PM_DATA_MANAGEMENT_PAGES = [
+  { path: 'load_data', component: LoadData },
+  { path: 'add_igv', component: AddIGV },
+]
+
 const DATA_MANAGEMENT_PAGES = [
+  ...PM_DATA_MANAGEMENT_PAGES,
   { path: 'sample_qc', component: SampleQc },
   { path: 'rna_seq', component: RnaSeq },
   { path: 'users', component: Users },
@@ -33,20 +41,23 @@ const ES_DATA_MANAGEMENT_PAGES = [
   ...DATA_MANAGEMENT_PAGES,
 ]
 
-const dataManagementPages = elasticsearchEnabled => (
-  elasticsearchEnabled ? ES_DATA_MANAGEMENT_PAGES : DATA_MANAGEMENT_PAGES
-)
+const dataManagementPages = (isDataManager, elasticsearchEnabled) => {
+  if (!isDataManager) {
+    return PM_DATA_MANAGEMENT_PAGES
+  }
+  return elasticsearchEnabled ? ES_DATA_MANAGEMENT_PAGES : DATA_MANAGEMENT_PAGES
+}
 
 const mapPageHeaderStateToProps = state => ({
-  pages: dataManagementPages(getElasticsearchEnabled(state)),
+  pages: dataManagementPages(getUser(state).isDataManager, getElasticsearchEnabled(state)),
 })
 
 export const DataManagementPageHeader = connect(mapPageHeaderStateToProps)(SimplePageHeader)
 
 const DataManagement = ({ match, user, elasticsearchEnabled }) => (
-  user.isDataManager ? (
+  (user.isDataManager || user.isPm) ? (
     <Switch>
-      {dataManagementPages(elasticsearchEnabled).map(({ path, params, component }) => (
+      {dataManagementPages(user.isDataManager, elasticsearchEnabled).map(({ path, params, component }) => (
         <Route key={path} path={`${match.url}/${path}${params || ''}`} component={component} />))}
       <Route exact path={match.url} component={null} />
       <Route component={Error404} />