+++ ui/pages/DataManagement/reducers.js
@@ -9,6 +9,7 @@ const RECEIVE_ELASTICSEARCH_STATUS = 'RECEIVE_ELASTICSEARCH_STATUS'
 const RECEIVE_PIPELINE_UPLOAD_STATS = 'RECEIVE_PIPELINE_UPLOAD_STATS'
 const RECEIVE_RNA_SEQ_UPLOAD_STATS = 'RECEIVE_RNA_SEQ_UPLOAD_STATS'
 const RECEIVE_PHE_PRI_UPLOAD_STATS = 'RECEIVE_PHE_PRI_UPLOAD_STATS'
+const RECEIVE_IGV_UPLOAD_STATS = 'RECEIVE_IGV_UPLOAD_STATS'
 const REQUEST_ALL_USERS = 'REQUEST_ALL_USERS'
 const RECEIVE_ALL_USERS = 'RECEIVE_ALL_USERS'
 
@@ -53,28 +54,48 @@ export const uploadQcPipelineOutput = values => submitRequest(
 
 export const deleteEsIndex = index => submitRequest('delete_index', RECEIVE_ELASTICSEARCH_STATUS, { index })
 
-export const uploadRnaSeq = values => (dispatch) => {
+const loadMultipleData = (path, getUpdateData, dispatchType, formatSuccessMessage) => values => (dispatch) => {
   let successResponseJson = null
   return new HttpRequestHelper(
-    '/api/data_management/update_rna_seq',
+    `/api/data_management/${path}`,
     (responseJson) => {
       successResponseJson = responseJson
     },
   ).post(values).then(() => {
-    const { info, warnings, sampleGuids, fileName } = successResponseJson
+    const { info, warnings } = successResponseJson
     let numLoaded = 0
-    return Promise.all(sampleGuids.map(sampleGuid => new HttpRequestHelper(
-      `/api/data_management/load_rna_seq_sample/${sampleGuid}`,
-      () => {
-        numLoaded += 1
-      },
-      e => warnings.push(`Error loading ${sampleGuid}: ${e.body && e.body.error ? e.body.error : e.message}`),
-    ).post({ fileName, dataType: values.dataType }))).then(() => {
-      info.push(`Successfully loaded data for ${numLoaded} RNA-seq samples`)
-      dispatch({ type: RECEIVE_RNA_SEQ_UPLOAD_STATS, newValue: { info, warnings } })
+    return Promise.all(getUpdateData(successResponseJson, values).map(
+      ([entityUrl, entityId, body]) => new HttpRequestHelper(
+        entityUrl,
+        () => {
+          numLoaded += 1
+        },
+        e => warnings.push(`Error loading ${entityId}: ${e.body && e.body.error ? e.body.error : e.message}`),
+      ).post(body),
+    )).then(() => {
+      info.push(formatSuccessMessage(numLoaded))
+      dispatch({ type: dispatchType, newValue: { info, warnings } })
     })
   })
 }
+
+export const uploadRnaSeq = loadMultipleData(
+  'update_rna_seq',
+  ({ sampleGuids, fileName }, { dataType }) => sampleGuids.map(sampleGuid => ([
+    `/api/data_management/load_rna_seq_sample/${sampleGuid}`, sampleGuid, { fileName, dataType },
+  ])),
+  RECEIVE_RNA_SEQ_UPLOAD_STATS,
+  numLoaded => `Successfully loaded data for ${numLoaded} RNA-seq samples`,
+)
+
+export const addIgv = loadMultipleData(
+  'add_igv',
+  ({ updates }) => updates.map(({ individualGuid, individualId, ...update }) => ([
+    `/api/individual/${individualGuid}/update_igv_sample`, individualId, update,
+  ])),
+  RECEIVE_IGV_UPLOAD_STATS,
+  numLoaded => `Successfully added IGV tracks for ${numLoaded} samples`,
+)
 
 export const uploadPhenotypePrioritization = values => submitRequest(
   'load_phenotype_prioritization_data', RECEIVE_PHE_PRI_UPLOAD_STATS, values,
@@ -86,6 +107,7 @@ export const reducers = {
   qcUploadStats: createSingleValueReducer(RECEIVE_PIPELINE_UPLOAD_STATS, {}),
   rnaSeqUploadStats: createSingleValueReducer(RECEIVE_RNA_SEQ_UPLOAD_STATS, {}),
   phePriUploadStats: createSingleValueReducer(RECEIVE_PHE_PRI_UPLOAD_STATS, {}),
+  igvUploadStats: createSingleValueReducer(RECEIVE_IGV_UPLOAD_STATS, {}),
   allUsers: createSingleValueReducer(RECEIVE_ALL_USERS, [], 'users'),
   allUsersLoading: loadingReducer(REQUEST_ALL_USERS, RECEIVE_ALL_USERS),
 }